Feature/pplus 1141 / divided interface (#42)
* PPLUS-1142: Console commands integrator logic / divided interfaces

* PPLUS-1141: divided interface / review changes

* PPLUS-1141: divided interface / review changes
