+++ src/Builder/ClassModifier/ClassConstant/Unwire/UnwireModifierInterface.php
@@ -1,0 +1,28 @@
+<?php
+
+/**
+ * Copyright © 2016-present Spryker Systems GmbH. All rights reserved.
+ * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
+ */
+
+declare(strict_types=1);
+
+namespace SprykerSdk\Integrator\Builder\ClassModifier\ClassConstant\Unwire;
+
+use SprykerSdk\Integrator\Transfer\ClassInformationTransfer;
+
+interface UnwireModifierInterface
+{
+    /**
+     * @param \SprykerSdk\Integrator\Transfer\ClassInformationTransfer $classInformationTransfer
+     * @param string $classNameToRemove
+     * @param string $targetMethodName
+     *
+     * @return \SprykerSdk\Integrator\Transfer\ClassInformationTransfer|null
+     */
+    public function unwire(
+        ClassInformationTransfer $classInformationTransfer,
+        string $classNameToRemove,
+        string $targetMethodName
+    ): ?ClassInformationTransfer;
+}
