+++ src/Builder/ClassModifier/ClassInstance/Unwire/UnwireModifier.php
@@ -7,20 +7,21 @@
 
 declare(strict_types=1);
 
-namespace SprykerSdk\Integrator\Builder\ClassModifier;
+namespace SprykerSdk\Integrator\Builder\ClassModifier\ClassInstance\Unwire;
 
-use RuntimeException;
 use SprykerSdk\Integrator\Builder\Checker\ClassMethodCheckerInterface;
+use SprykerSdk\Integrator\Builder\ClassModifier\AddVisitorsTrait;
+use SprykerSdk\Integrator\Builder\ClassModifier\CommonClass\CommonClassModifierInterface;
 use SprykerSdk\Integrator\Builder\Finder\ClassNodeFinderInterface;
 use SprykerSdk\Integrator\Transfer\ClassInformationTransfer;
 use SprykerSdk\Integrator\Transfer\ClassMetadataTransfer;
 
-class ClassInstanceClassModifier implements ClassInstanceClassModifierInterface
+class UnwireModifier implements UnwireModifierInterface
 {
     use AddVisitorsTrait;
 
     /**
-     * @var \SprykerSdk\Integrator\Builder\ClassModifier\CommonClassModifierInterface
+     * @var \SprykerSdk\Integrator\Builder\ClassModifier\CommonClass\CommonClassModifierInterface
      */
     protected $commonClassModifier;
 
@@ -35,15 +36,15 @@ class ClassInstanceClassModifier implements ClassInstanceClassModifierInterface
     protected $classMethodChecker;
 
     /**
-     * @var array<\SprykerSdk\Integrator\Builder\ClassModifier\ClassInstanceModifierStrategy\ClassInstanceModifierStrategyInterface>
+     * @var array<\SprykerSdk\Integrator\Builder\ClassModifier\ClassInstanceModifierStrategy\Unwire\UnwireModifierStrategyInterface>
      */
     protected $classInstanceModifierStrategies;
 
     /**
-     * @param \SprykerSdk\Integrator\Builder\ClassModifier\CommonClassModifierInterface $commonClassModifier
+     * @param \SprykerSdk\Integrator\Builder\ClassModifier\CommonClass\CommonClassModifierInterface $commonClassModifier
      * @param \SprykerSdk\Integrator\Builder\Finder\ClassNodeFinderInterface $classNodeFinder
      * @param \SprykerSdk\Integrator\Builder\Checker\ClassMethodCheckerInterface $classMethodChecker
-     * @param array<\SprykerSdk\Integrator\Builder\ClassModifier\ClassInstanceModifierStrategy\ClassInstanceModifierStrategyInterface> $classInstanceModifierStrategies
+     * @param array<\SprykerSdk\Integrator\Builder\ClassModifier\ClassInstanceModifierStrategy\Unwire\UnwireModifierStrategyInterface> $classInstanceModifierStrategies
      */
     public function __construct(
         CommonClassModifierInterface $commonClassModifier,
@@ -61,41 +62,9 @@ class ClassInstanceClassModifier implements ClassInstanceClassModifierInterface
      * @param \SprykerSdk\Integrator\Transfer\ClassInformationTransfer $classInformationTransfer
      * @param \SprykerSdk\Integrator\Transfer\ClassMetadataTransfer $classMetadataTransfer
      *
-     * @throws \RuntimeException
-     *
-     * @return \SprykerSdk\Integrator\Transfer\ClassInformationTransfer
-     */
-    public function wireClassInstance(
-        ClassInformationTransfer $classInformationTransfer,
-        ClassMetadataTransfer $classMetadataTransfer
-    ): ClassInformationTransfer {
-        $targetMethodName = $classMetadataTransfer->getTargetMethodNameOrFail();
-        $methodNode = $this->classNodeFinder->findMethodNode($classInformationTransfer, $targetMethodName);
-        if (!$methodNode) {
-            $classInformationTransfer = $this->commonClassModifier->overrideMethodFromParent($classInformationTransfer, $targetMethodName);
-            $methodNode = $this->classNodeFinder->findMethodNode($classInformationTransfer, $targetMethodName);
-        }
-
-        if ($methodNode === null) {
-            throw new RuntimeException('Method node not found');
-        }
-
-        foreach ($this->classInstanceModifierStrategies as $classInstanceModifierStrategy) {
-            if ($classInstanceModifierStrategy->isApplicable($methodNode)) {
-                return $classInstanceModifierStrategy->wireClassInstance($classInformationTransfer, $classMetadataTransfer);
-            }
-        }
-
-        return $classInformationTransfer;
-    }
-
-    /**
-     * @param \SprykerSdk\Integrator\Transfer\ClassInformationTransfer $classInformationTransfer
-     * @param \SprykerSdk\Integrator\Transfer\ClassMetadataTransfer $classMetadataTransfer
-     *
      * @return \SprykerSdk\Integrator\Transfer\ClassInformationTransfer|null
      */
-    public function unwireClassInstance(
+    public function unwire(
         ClassInformationTransfer $classInformationTransfer,
         ClassMetadataTransfer $classMetadataTransfer
     ): ?ClassInformationTransfer {
