--- src/Builder/ClassModifier/ClassInstanceModifierStrategy/ClassInstanceModifierStrategyInterface.php
@@ -1,44 +1,0 @@
-<?php
-
-/**
- * Copyright © 2016-present Spryker Systems GmbH. All rights reserved.
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
- */
-
-namespace SprykerSdk\Integrator\Builder\ClassModifier\ClassInstanceModifierStrategy;
-
-use PhpParser\Node\Stmt\ClassMethod;
-use SprykerSdk\Integrator\Transfer\ClassInformationTransfer;
-use SprykerSdk\Integrator\Transfer\ClassMetadataTransfer;
-
-interface ClassInstanceModifierStrategyInterface
-{
-    /**
-     * @param \PhpParser\Node\Stmt\ClassMethod $node
-     *
-     * @return bool
-     */
-    public function isApplicable(ClassMethod $node): bool;
-
-    /**
-     * @param \SprykerSdk\Integrator\Transfer\ClassInformationTransfer $classInformationTransfer
-     * @param \SprykerSdk\Integrator\Transfer\ClassMetadataTransfer $classMetadataTransfer
-     *
-     * @return \SprykerSdk\Integrator\Transfer\ClassInformationTransfer
-     */
-    public function wireClassInstance(
-        ClassInformationTransfer $classInformationTransfer,
-        ClassMetadataTransfer $classMetadataTransfer
-    ): ClassInformationTransfer;
-
-    /**
-     * @param \SprykerSdk\Integrator\Transfer\ClassInformationTransfer $classInformationTransfer
-     * @param \SprykerSdk\Integrator\Transfer\ClassMetadataTransfer $classMetadataTransfer
-     *
-     * @return \SprykerSdk\Integrator\Transfer\ClassInformationTransfer
-     */
-    public function unwireClassInstance(
-        ClassInformationTransfer $classInformationTransfer,
-        ClassMetadataTransfer $classMetadataTransfer
-    ): ClassInformationTransfer;
-}
