+++ src/Builder/ClassModifier/GlueRelationship/Wire/WireModifierInterface.php
@@ -1,0 +1,32 @@
+<?php
+
+/**
+ * Copyright © 2016-present Spryker Systems GmbH. All rights reserved.
+ * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
+ */
+
+declare(strict_types=1);
+
+namespace SprykerSdk\Integrator\Builder\ClassModifier\GlueRelationship\Wire;
+
+use SprykerSdk\Integrator\Transfer\ClassInformationTransfer;
+
+interface WireModifierInterface
+{
+    /**
+     * @param \SprykerSdk\Integrator\Transfer\ClassInformationTransfer $classInformationTransfer
+     * @param string $targetMethodName
+     * @param string $key
+     * @param string $classNameToAdd
+     *
+     * @throws \RuntimeException
+     *
+     * @return \SprykerSdk\Integrator\Transfer\ClassInformationTransfer
+     */
+    public function wire(
+        ClassInformationTransfer $classInformationTransfer,
+        string $targetMethodName,
+        string $key,
+        string $classNameToAdd
+    ): ClassInformationTransfer;
+}
